+++ changelog.md
@@ -15,5 +15,6 @@
 * includeJssp 改为 this.includeJssp
 * D2JS.DataTable.nest(...) 改为 nest(d2js,...) 
 * 不再支持在 jssp 中使用 [% console.log() %]
+* $V(rcd, validators) 改为 $V(d2js, rcd, validators), 校验器规格随之调整，check 函数由 function(v, fld, rcd) 扩充为 function(v, fld, rcd, d2js)
 
 本轮优化后，d2js 性能大致接近 jsp 的 50% - 90%。